squash! b387a4a6f9f9b3194a29c1a80c39c983d5dc4fd5
fil_node_t::rename(), fil_node_t::set_backup_name(),
fil_node_t::get_backup_name(), fil_node_t::~fil_node_t():
New code, to allow InnoDB_backup::init() get a
consistent snapshot of the file names, preserved in
fil_node_t::backup_name. If files that were created
before the latest LSN are renamed during the backup,
we will copy them using the old name and rely on the
application of FILE_RENAME records.

+++ storage/innobase/os/os0file.cc
@@ -3616,7 +3616,7 @@ void fil_node_t::find_metadata(IF_WIN(,bool create)) noexcept
   struct stat statbuf;
   if (!fstat(file, &statbuf))
   {
-    block_size= statbuf.st_blksize;
+    block_size= uint32_t(statbuf.st_blksize);
 # ifdef __linux__
     on_ssd= fil_system.is_ssd(statbuf.st_dev);
 # endif
